style: -Info 클래스 하위 패키지로 이동
