+++ src/main/java/com/wanted/external/info/impl/TwitterApiInfo.java
@@ -1,5 +1,7 @@
-package com.wanted.external.info;
+package com.wanted.external.info.impl;
 
+import com.wanted.external.info.ApiSpec;
+import com.wanted.external.info.SnsApiInfo;
 import java.util.Optional;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
@@ -8,13 +10,13 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.MultiValueMap;
 
 @Component
-public class ThreadApiInfo implements SnsApiInfo {
+public class TwitterApiInfo implements SnsApiInfo {
 
     @Getter
-    @Value("${thread.api.endpoint:}")
+    @Value("${twitter.api.endpoint:}")
     private String endpoint;
 
-    @Value("#{${thread.api.keys:{ : }}}")
+    @Value("#{${twitter.api.keys:{ : }}}")
     private MultiValueMap<String, String> keys;
 
     @Override
